board: sue: s810: take first found PHY address for FEC2 and patch dtb
This will `scan` the MDIO bus in the range 0-7, take the first found
phy address and _assume_ it is connected to the internal MAC (fec2)
on the module. Before booting the kernel the fec2 devicetree node
wil be patched accordingly.

We need this because the ADC values are the same for both modules
but the PHY addresses changed from one revision to another.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>