board: sue: kobs: fixup bad block marker
The ROM-bootloader expects the FCB to be `randomized`, in the `kobs-ng`
userspace tool, the randomization is done in software utilizing a very
large array and some logic which represents the HW randomizer. Since we
do not want to include this large array in our u-boot binary, we are
enabling the HW randomizer before writing the FCB. The problem is, that
the bad block marker (which should be 0xFF for good blocks) will also be
randomized and therefore not 0xFF anymore. This creates some issues,
especially for the userspace, since we cannot write using `kobs-ng` to these
blocks anymore. As a workaround we write the page with the HW randomizer
enabled, then we read the page back, set the bad block marker to 0xFF and
write back the new page.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>