board: sue: stream810: patch operating points if desired by const switch
This allows us to specify `additional-operating-points` in the devicetree,
which will be enabled when the `enable_additional_operating_points`
variable in the const partition is set.

We use this mechanism to allow customers to enable or disable 1.2GHz
operation of the CPU.

Essentially we are just taking the data from the
`additional-operating-points` property and append it to the current
`operating-points` before booting.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>