sue: device_interface: add support daughter board and StreamKitPrime
We extend the functionality of the HW detection for daughterboards. This
is needed for StreamKitPrime since the LSB on the StreamKitPrime contains
the ADC value from the daughter board.

In addition to that we store the raw ADC values and the codes in the
device_info struct to be able to access them for printing and debugging
purposes.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>