configs: sue_fwupdate: check fit and u-boot-env partitions when determining factory state
Swupdate performs scanning for UBI volumes while it's running. This process
involves attaching the UBI volume. If an empty mtd partition is detected,
it is first formatted and then attached. This can cause troubles while
the board is in factory state: if it was rebooted (without anything
written to the flash), u-boot would no longer think the board is in
factory state.

A solution is to check if the u-boot-env partition is empty instead.

Signed-off-by: Maros Medvec <maros.medvec@streamunlimited.com>