board: sue: s810: imximage_1x2Gb: configure refresh timing to 1x4Gb
We configure the refresh timing of the 1x2Gb config to the same value
as the 1x4Gb config. This allows us to use the the 1x2Gb config for the
1x4Gb chip in the factory.

It was confirmed by the HW department and by NANYA that this value is fine
for the 1x2Gb chip.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>