board: sue: device_interface: Use generic_board_ops for everything except DemoClient
It was a wish from the HW department to be able to trigger the software
update on the factory tester and an unknown board as well. Thus we
created the generic_board_ops which only check the NPB_IN and used these
generic_board_ops for everything which is not the DemoClient.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>